+++ server.php
@@ -68,13 +68,19 @@ $server = new Worker( "Memcached_Text://127.0.0.1:{$options['port']}" );
 $server->count = 4;
 $server->name = 'Marmerine v'.MARMERINE_VERSION;
 
+$server->onWorkerStart = static function() {
+	global $storage;
+	//$storage = new Memcached_Storage( ':memory:' );
+	$storage = new Memcached_Storage( __DIR__ . '/data/marmerine.db' );
+};
+
 $server->onConnect = static function ( TcpConnection $conn ) {
 	bump_stat( 'total_connections' );
 };
 
 $server->onMessage = static function ( TcpConnection $conn, object $data ) {
-#	$storage = new Memcached_Storage( ':memory:' );
-	$storage = new Memcached_Storage( __DIR__ . '/data/marmerine.db' );
+
+	global $storage;
 
 	bump_stat( "cmd_{$data->command}" );