+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [#5](https://github.com/josephscott/marmerine/pull/5) @joanhey - Use static anonymous functions, provides a small memory improvement
 - [#8](https://github.com/josephscott/marmerine/pull/8) @joanhey - Set the workerman name
 - [#7](https://github.com/josephscott/marmerine/pull/7) @joanhey - Remove int casting from verbose()
+- [#9](https://github.com/josephscott/marmerine/pull/9) @joanhey - Test file naming style changes
 
 
 ## 0.0.3 - 2022-05-21