+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Set a busy timeout for the SQLite database, and always enable WAL
 - [#12](https://github.com/josephscott/marmerine/pull/12) @joanhey - Send ERROR when a command is not supported
 - [#16](https://github.com/josephscott/marmerine/pull/16) @joanhey - Add Github action 
+- [#16](https://github.com/josephscott/marmerine/pull/16) @joanhey - Add ENV support for setting the port of the server to run tests against 
 
 
 ## 0.0.3 - 2022-05-21